+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Log object ID if poll data is not valid.
+- Log database errors coming from inbox handler with level ERROR.
 
 ## [3.13.1] - 2025-01-03